remove Bearer/SSO
