+++ src/main/java/com/owncloud/android/lib/common/OwnCloudCredentialsFactory.java
@@ -25,7 +25,6 @@
 package com.owncloud.android.lib.common;
 
 public class OwnCloudCredentialsFactory {
-
 	public static final String CREDENTIAL_CHARSET = "UTF-8";
 
 	private static OwnCloudAnonymousCredentials sAnonymousCredentials;
@@ -34,47 +33,10 @@ public class OwnCloudCredentialsFactory {
 		return new OwnCloudBasicCredentials(username, password);
 	}
 
-    public static OwnCloudCredentials newBearerCredentials(String authToken) {
-        return new OwnCloudBearerCredentials(authToken);
-	}
-    
-	public static OwnCloudCredentials newSamlSsoCredentials(String username, String sessionCookie) {
-		return new OwnCloudSamlSsoCredentials(username, sessionCookie);
-	}
-
 	public static final OwnCloudCredentials getAnonymousCredentials() {
 		if (sAnonymousCredentials == null) {
 			sAnonymousCredentials = new OwnCloudAnonymousCredentials();
 		}
 		return sAnonymousCredentials;
 	}
-
-	public static final class OwnCloudAnonymousCredentials implements OwnCloudCredentials {
-		
-		protected OwnCloudAnonymousCredentials() {
-		}
-		
-		@Override
-		public void applyTo(OwnCloudClient client) {
-			client.getState().clearCredentials();
-			client.getState().clearCookies();
-		}
-
-		@Override
-		public String getAuthToken() {
-			return "";
-		}
-
-		@Override
-		public boolean authTokenExpires() {
-			return false;
-		}
-
-		@Override
-		public String getUsername() {
-			// no user name
-			return null;
-		}
-	}
-
 }
