+++ src/main/java/com/owncloud/android/lib/common/accounts/AccountUtils.java
@@ -50,7 +50,6 @@ public class AccountUtils {
 	public static final String WEBDAV_PATH_4_0 = "/remote.php/webdav";
 	public static final String WEBDAV_PATH_9_0 = "/remote.php/dav";
     public static final String DAV_UPLOAD = "/remote.php/dav/uploads";
-	public static final String ODAV_PATH = "/remote.php/odav";
     public static final String STATUS_PATH = "/status.php";
 
     /**
